SRC/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
pillow
opencv-python
numpy
python-dotenv
pytest
httpx
pytesseract  # For OCR in ERD parsing
